Summary: Fixing a bug, the VMs constructor should have taken a double value not a float. Also making EnergyUsageSource abstract.

+++ iaas/energy-modeller/src/main/java/eu/ascetic/asceticarchitecture/iaas/energymodeller/types/energyuser/VM.java
@@ -51,7 +51,7 @@ public class VM extends EnergyUsageSource {
      * @param ramMb The amount of RAM in MB.
      * @param diskGb The size of the disk in GB.
      */
-    public VM(int cpus, int ramMb, float diskGb) {
+    public VM(int cpus, int ramMb, double diskGb) {
         this.cpus = cpus;
         this.ramMb = ramMb;
         this.diskGb = diskGb;
